requirements.txt
requests
jsonpath-ng
python-dotenv
datetime
schedule